% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tuto.R
\name{check_tuto}
\alias{check_tuto}
\title{Check the code/solutions of a tutorial}
\usage{
check_tuto(paths = NULL)
}
\arguments{
\item{paths}{A vector containing tutorial paths.}

\item{file}{A connection, or a character string naming the file to print to. If "" (the default), prints to the standard output connection.}
}
\description{
Check the code/solutions of a tutorial
}
\examples{
tuto <- Sys.glob(file.path(system.file(package="rtrainer"), "tutorials", "*", "*.Rmd"))
check_tuto(tuto)
}
