% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_text.R
\name{extract_text}
\alias{extract_text}
\title{Delete the code from the tutorials and extract the text.}
\usage{
extract_text(paths = NULL, file = "")
}
\arguments{
\item{paths}{A vector containing tutorial paths.}

\item{file}{A connection, or a character string naming the file to print to. If "" (the default), prints to the standard output connection.}
}
\description{
Delete the code from the tutorials and extract the text.
}
\examples{
tuto <- Sys.glob(file.path(system.file(package="rtrainer"), "tutorials", "*", "*.Rmd"))
out <- tempfile()
extract_text(tuto, file=out)
}
