% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{gtts_yearly}
\alias{gtts_yearly}
\title{The gtts_yearly dataset
Global Temperature Time Series. Data are included from the
GISS Surface Temperature (GISTEMP) analysis and the global component
of Climate at a Glance (GCAG). Two datasets are provided: 1) global
monthly mean and 2) annual mean temperature anomalies in degrees
Celsius from 1880 to the present.}
\format{
A data frame with 274 rows and 3 variables
}
\usage{
data(gtts_yearly)
}
\description{
Global temperature anomaly data come from the Global Historical Climatology Network-Monthly
(GHCN-M) data set and International Comprehensive Ocean-Atmosphere Data Set (ICOADS),
which have data from 1880 to the present. These two datasets are blended into a single
product to produce the combined global land and ocean temperature anomalies.
The available timeseries of global-scale temperature anomalies are calculated with respect
to the 20th century average
}
\details{
Obtained from:  https://data.giss.nasa.gov/gistemp/

1/ GISTEMP: NASA Goddard Institute for Space Studies (GISS) Surface Temperature Analysis, Global
Land-Ocean Temperature Index.
2/ NOAA National Climatic Data Center (NCDC), global component of Climate at a Glance (GCAG).

Downloaded from: https://datahub.io/core/global-temp

Dataset information:
·        Number of rows: 274
·        Number of columns: 3

\itemize{
\item Source: Global Land-Ocean Temperature Index (GISTEMP); Global component of Climate at a Glance (GCAG)
\item Year: integer
\item Mean: The available timeseries of global-scale temperature anomalies are calculated with respect
to the 20th century average
}
}
\keyword{datasets}
