% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_fasta_fragilaria}
\alias{write_fasta_fragilaria}
\title{Writes the fragilaria dataset to files}
\usage{
write_fasta_fragilaria(dir_path = NULL)
}
\value{
The paths to the directory where file have been stored.
}
\description{
Writes the fragilaria dataset to files
}
\examples{
write_fasta_fragilaria()
}
